refactor(settlement-service): move the new admin tests to tests/admin.rs
